src/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
distro==1.9.0
Django==5.0.6
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
idna==3.7
netaddr==1.3.0
pillow==10.4.0
requests==2.32.3
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yookassa==3.3.0
